pyproject.toml
[project]
name = "autoseo"
version = "1.0.0"
description = "Automated SEO content generation pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
autoseo = "run:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["pipeline"]
omit = ["tests/*"]
